cpufreq  GTK3: hardcode fewer labels
In gtk3,  replacing "label" with "applet->label" makes the gtk_widget_get_preferred_width functions work for the frequency and unit labels without issue. Trim them a bit in gtk3.19 builds as they render 2px wider there for some reason. The exception is the percentage label. That one apparently always starts with "0%" entered-and it does NOT update it's width when larger character strings come up. Thus, a fixed figure of 20px still has to be added to that one. On the other hand, that will ensure the applet never jumps from percentage changes. Height of the whole applet is still limited somewhere to 24 px, still looking for that issue as it hurts the match to taller panels.