CpuFreq: fix warning from GTK_ORIENTABLE
gtk_orientable_get_orientation now throws warnings so remove it. It does not appear to work anyway.  Applet tested in a vertical (left side) panel as well as a horizontal one, as long as the vertical panel is wide enough to accept the applet it works and puts the frequency and units on two levels. Puts them on one level in a horizontal panel.