Force immediate rendering of icon/label/units as appropriate
Whichever of the icons and labels are configured to be visible are
forced to be rendered at applet instantiation instead of waiting for the
monitor to run. "---" and "--" are used for the labels so as not to
display wrong information.

Also save and restore any existing label strings when recomputing widths
to avoid showing the user wrong strings left over from running the max
width routines (avoids the side effects noted by @monsta in #268).