Add cpupower-4.7 compatibility
Drops linux version detection as that relies on linux headers version,
not actual cpupower version (they may be independent, as with Gentoo)

Instead, detect presence of cpupower.h header file.

Starting with cpupower-4.9 in Gentoo, we are installing this header even
though it has not been resolved yet in upstream kernel's cpupower
installer.  Mike Gilbert (floppym@gentoo.org) has told me that he
intends to email the linux-pm mailing list about it in the near future.

This is probably the best solution, assuming that other distros have
this header installed (and eventually that upstream will)