invest: Bring back the invest applet
The original invest applet was dropped in
https://github.com/mate-desktop/mate-applets/commit/41f7b7a2a38ca9f3929fcbd88d228b3815034876
due to a broken API. This brings it back using the modern JSON version
of the Yahoo Finance API, a rewrite from Python to C, and a simplified
interface (i.e. instead of the full portfolio management app,
I simplified it into a stock ticker price tracker).